Organize the WASI submodules for downstream use
In particular this changes the module structure so the implementations
may be used without the generated interfaces.
