--- brevity-wasi-main/src/jvmMain/kotlin/brevity/wasi/RealRandom.kt
@@ -1,40 +1,0 @@
-package brevity.wasi
-
-import java.security.SecureRandom
-import kotlin.random.Random
-import kotlin.random.asKotlinRandom
-import kotlin.random.nextULong
-import okio.ByteString
-import okio.ByteString.Companion.toByteString
-import wit.wasi.random.v0_3_0.Insecure
-import wit.wasi.random.v0_3_0.InsecureSeed
-import wit.wasi.random.v0_3_0.Random as wasiRandom
-
-class RealRandom : wasiRandom {
-  val secureRandom = SecureRandom().asKotlinRandom()
-
-  override fun getRandomBytes(maxLen: ULong): ByteString {
-    return secureRandom.nextBytes(maxLen.toInt()).toByteString()
-  }
-
-  override fun getRandomU64(): ULong {
-    return secureRandom.nextULong()
-  }
-}
-
-class RealInsecureRandom : Insecure {
-  override fun getInsecureRandomBytes(maxLen: ULong): ByteString {
-    return Random.nextBytes(maxLen.toInt()).toByteString()
-  }
-
-  override fun getInsecureRandomU64(): ULong {
-    return Random.nextULong()
-  }
-
-}
-
-class RealInsecureSeed : InsecureSeed {
-  override fun getInsecureSeed(): Pair<ULong, ULong> {
-    return Pair(Random.nextULong(), Random.nextULong())
-  }
-}
