--- brevity-wasi-p2/src/jvmMain/kotlin/brevity/wasi/p2/RealEnvironment.kt
@@ -1,17 +1,0 @@
-package brevity.wasi.p2
-
-import wit.wasi.cli.v0_2_0.Environment
-
-class RealEnvironment : Environment {
-  override fun getEnvironment(): List<Pair<String, String>> {
-    return listOf()
-  }
-
-  override fun getArguments(): List<String> {
-    return listOf()
-  }
-
-  override fun initialCwd(): String? {
-    return null
-  }
-}
