--- brevity-wasi-p2/src/jvmMain/kotlin/brevity/wasi/p2/RealPollable.kt
@@ -1,12 +1,0 @@
-package brevity.wasi.p2
-
-import wit.wasi.io.v0_2_0.Poll
-
-class RealPollable : Poll.Pollable {
-  override fun ready(): Boolean {
-    return true
-  }
-
-  override fun block() {
-  }
-}
