+++ settings.gradle.kts
@@ -21,9 +21,10 @@ include(":brevity-gradle-plugin")
 include(":brevity-integration-tests")
 include(":brevity-kotlin-generator")
 include(":brevity-testing")
-include(":brevity-wasi-main")
-include(":brevity-wasi-p1")
-include(":brevity-wasi-p2")
 include(":brevity-wit")
+include(":wasi:brevity-wasi")
+include(":wasi:brevity-wasi-p1")
+include(":wasi:brevity-wasi-p2")
+include(":wasi:brevity-wasi-p3")
 
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
