+++ wasi/brevity-wasi-p1/src/commonMain/kotlin/wit/wasi/v0_1/Fds.kt
