+++ wasi/brevity-wasi/README.md
@@ -1,0 +1,7 @@
+Brevity’s WASI Implementation
+=============================
+
+This is a Kotlin implementation of the WASI host APIs, intended for use in whole or in part.
+
+Note that this package has a `compileOnly` dependency on Brevity's generated WASI API modules. Users
+may choose to use these implementations with host APIs generated elsewhere.
