+++ wasi/brevity-wasi/src/jvmMain/kotlin/dev/wasmo/brevity/wasi/p2/RealPollable.kt
@@ -1,0 +1,12 @@
+package dev.wasmo.brevity.wasi.p2
+
+import wit.wasi.io.v0_2_0.Poll
+
+class RealPollable : Poll.Pollable {
+  override fun ready(): Boolean {
+    return true
+  }
+
+  override fun block() {
+  }
+}
